Relocate code to new module Signing
